% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNRCINT.R
\name{urootp}
\alias{urootp}
\title{Title Unit Root Testing}
\usage{
urootp(x)
}
\arguments{
\item{x}{Only Numeric values}
}
\value{
Returns a Data frame of unit Root test with Critical and P-Values
}
\description{
Title Unit Root Testing
}
\examples{
urootp(data)
}
